split scanner_window & tag_dispatcher into separate files
